build(infra): move all non-stack code to dedicated folder
Ref: metriport/metriport-internal#1050
